requirements.txt
Django==2.1.3
django-rest-framework==0.1.0
Pillow==5.3.0
factory-boy==2.11.1
django-simple-history==2.5.1
